+++ app/sprinkles/core/src/Bakery/MigrateCommand.php
@@ -11,8 +11,8 @@ namespace UserFrosting\Sprinkle\Core\Bakery;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Input\InputOption;
+use UserFrosting\Sprinkle\Core\Bakery\Helper\ConfirmableTrait;
 use UserFrosting\System\Bakery\BaseCommand;
-use UserFrosting\System\Bakery\ConfirmableTrait;
 
 /**
  * migrate Bakery Command